+++ src/app/user/user-database/user-database.component.ts
@@ -41,8 +41,9 @@ export class UserDatabaseComponent implements OnInit {
 
      // maxWidth:120,
       sorting: true,
-      suppressMenu: true,
+     // suppressMenu: true,
       sortable: true,
+      filter: true,
       cellStyle: function (params) {
         return {
           cursor: 'pointer',