+++ src/app/xrd/xrd-index/xrd-index.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-xrd-index',
+  templateUrl: './xrd-index.component.html',
+  styleUrls: ['./xrd-index.component.css']
+})
+export class XrdIndexComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}