+++ src/app/xrd/xrd-view/xrd-view.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-xrd-view',
+  templateUrl: './xrd-view.component.html',
+  styleUrls: ['./xrd-view.component.css']
+})
+export class XrdViewComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}